+++ test/smokes/rubocop/unexpected_error/backtrace_silencers.rb
@@ -1,0 +1,3 @@
+# https://github.com/rubocop-hq/rubocop/issues/8674
+
+Rails.backtrace_cleaner.remove_silencers! if ENV['CI']